Fixed: flower firing once then doing nothing
It was due to the firing code being executed too close to the start time.
So to be sure it is executed in the update function (at least to begin)
we wait for the end of the frame before doing the commands.
It also works with waiting the fixed update but I was not sure so oh well...